# server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/warehouse/category/list": {
      "get": {
        "operationId": "CategoryController_findAll",
        "summary": "查询分类列表",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "分类名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资分类管理"
        ]
      }
    },
...
    "/warehouse/category/{id}": {
      "get": {
        "operationId": "CategoryController_findOne",
        "summary": "查询分类详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资分类管理"
        ]
      },
      "delete": {
        "operationId": "CategoryController_remove",
        "summary": "删除分类",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资分类管理"
        ]
      }
    },
    "/warehouse/category": {
      "post": {
        "operationId": "CategoryController_create",
        "summary": "新增分类",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "物资分类管理"
        ]
      },
      "put": {
        "operationId": "CategoryController_update",
        "summary": "修改分类",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资分类管理"
        ]
      }
    },
    "/warehouse/supplier/list": {
      "get": {
        "operationId": "SupplierController_findAll",
        "summary": "查询供应商列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "供应商名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contact",
            "required": false,
            "in": "query",
            "description": "联系人",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "供应商管理"
        ]
      }
    },
...
    "/warehouse/supplier/{id}": {
      "get": {
        "operationId": "SupplierController_findOne",
        "summary": "查询供应商详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "供应商管理"
        ]
      },
      "delete": {
        "operationId": "SupplierController_remove",
        "summary": "删除供应商",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "供应商管理"
        ]
      }
    },
    "/warehouse/supplier": {
      "post": {
        "operationId": "SupplierController_create",
        "summary": "新增供应商",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "供应商管理"
        ]
      },
      "put": {
        "operationId": "SupplierController_update",
        "summary": "修改供应商",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "供应商管理"
        ]
      }
    },
    "/warehouse/material/list": {
      "get": {
        "operationId": "MaterialController_findAll",
        "summary": "查询物资列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "物资名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "物资编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "分类ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资管理"
        ]
      }
    },
...
    "/warehouse/material/{id}": {
      "get": {
        "operationId": "MaterialController_findOne",
        "summary": "查询物资详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资管理"
        ]
      },
      "delete": {
        "operationId": "MaterialController_remove",
        "summary": "删除物资",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资管理"
        ]
      }
    },
    "/warehouse/material": {
      "post": {
        "operationId": "MaterialController_create",
        "summary": "新增物资",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaterialDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "物资管理"
        ]
      },
      "put": {
        "operationId": "MaterialController_update",
        "summary": "修改物资",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMaterialDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "物资管理"
        ]
      }
    },
    "/warehouse/stock/record/list": {
      "get": {
        "operationId": "StockController_findAll",
        "summary": "查询出入库记录列表",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "materialName",
            "required": false,
            "in": "query",
            "description": "物资名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "操作类型：in入库 out出库",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "beginTime",
            "required": false,
            "in": "query",
            "description": "开始时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endTime",
            "required": false,
            "in": "query",
            "description": "结束时间",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "出入库管理"
        ]
      }
    },
    "/warehouse/stock/record/{id}": {
      "get": {
        "operationId": "StockController_findOne",
        "summary": "查询出入库记录详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "出入库管理"
        ]
      }
    },
    "/warehouse/stock/in": {
      "post": {
        "operationId": "StockController_stockIn",
        "summary": "入库操作",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "出入库管理"
        ]
      }
    },
    "/warehouse/stock/out": {
      "post": {
        "operationId": "StockController_stockOut",
        "summary": "出库操作",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockOutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "出入库管理"
        ]
      }
    },
    "/warehouse/stock/recent": {
      "get": {
        "operationId": "StockController_getRecentRecords",
        "summary": "获取最近出入库记录",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "出入库管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "code": {
            "type": "string",
            "description": "分类编码"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID"
          },
          "orderNum": {
            "type": "number",
            "description": "显示顺序"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "code": {
            "type": "string",
            "description": "分类编码"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID"
          },
          "orderNum": {
            "type": "number",
            "description": "显示顺序"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID"
          }
        },
        "required": [
          "name",
          "code",
          "categoryId"
        ]
      },
      "CreateSupplierDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "供应商编码"
          },
          "name": {
            "type": "string",
            "description": "供应商名称"
          },
          "contact": {
            "type": "string",
            "description": "联系人"
          },
          "phone": {
            "type": "string",
            "description": "联系电话"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "address": {
            "type": "string",
            "description": "地址"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": [
          "code",
          "name",
          "contact",
          "phone"
        ]
      },
      "UpdateSupplierDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "供应商编码"
          },
          "name": {
            "type": "string",
            "description": "供应商名称"
          },
          "contact": {
            "type": "string",
            "description": "联系人"
          },
          "phone": {
            "type": "string",
            "description": "联系电话"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "address": {
            "type": "string",
            "description": "地址"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "supplierId": {
            "type": "number",
            "description": "供应商ID"
          }
        },
        "required": [
          "code",
          "name",
          "contact",
          "phone",
          "supplierId"
        ]
      },
      "CreateMaterialDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "物资编码"
          },
          "name": {
            "type": "string",
            "description": "物资名称"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "specification": {
            "type": "string",
            "description": "规格型号"
          },
          "unit": {
            "type": "string",
            "description": "单位"
          },
          "safeStock": {
            "type": "number",
            "description": "安全库存"
          },
          "supplierId": {
            "type": "number",
            "description": "供应商ID"
          },
          "supplierName": {
            "type": "string",
            "description": "供应商名称"
          },
          "price": {
            "type": "number",
            "description": "单价"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": [
          "code",
          "name",
          "categoryId",
          "unit"
        ]
      },
      "UpdateMaterialDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "物资编码"
          },
          "name": {
            "type": "string",
            "description": "物资名称"
          },
          "categoryId": {
            "type": "number",
            "description": "分类ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "specification": {
            "type": "string",
            "description": "规格型号"
          },
          "unit": {
            "type": "string",
            "description": "单位"
          },
          "safeStock": {
            "type": "number",
            "description": "安全库存"
          },
          "supplierId": {
            "type": "number",
            "description": "供应商ID"
          },
          "supplierName": {
            "type": "string",
            "description": "供应商名称"
          },
          "price": {
            "type": "number",
            "description": "单价"
          },
          "status": {
            "type": "string",
            "description": "状态"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "materialId": {
            "type": "number",
            "description": "物资ID"
          }
        },
        "required": [
          "code",
          "name",
          "categoryId",
          "unit",
          "materialId"
        ]
      },
      "StockInDto": {
        "type": "object",
        "properties": {
          "materialId": {
            "type": "number",
            "description": "物资ID"
          },
          "quantity": {
            "type": "number",
            "description": "入库数量"
          },
          "supplierId": {
            "type": "number",
            "description": "供应商ID"
          },
          "supplierName": {
            "type": "string",
            "description": "供应商名称"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": [
          "materialId",
          "quantity"
        ]
      },
      "StockOutDto": {
        "type": "object",
        "properties": {
          "materialId": {
            "type": "number",
            "description": "物资ID"
          },
          "quantity": {
            "type": "number",
            "description": "出库数量"
          },
          "receiver": {
            "type": "string",
            "description": "领用人"
          },
          "department": {
            "type": "string",
            "description": "领用部门"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          }
        },
        "required": [
          "materialId",
          "quantity",
          "receiver"
        ]
      }
    }
  }
}